Object Oriented Programming/Math_painting/requirements.txt
autopep8==1.6.0
numpy==1.21.5
Pillow==8.4.0
pycodestyle==2.8.0
toml==0.10.2